feat: fraction CompanyReviewModal and Tests
